mm: compaction: Add automatic compaction mechanism
This adds a simple implementation of a scheduled compaction
mechanism.

The scheduled compaction gets executed compaction_soff_delay_ms
(default 3000ms) after the screen gets turned off unless the
compaction_timeout_ms (default 900000ms) prevents it because
a memory compaction is not required.

The workqueue is freezable to indicate that this shouldn't be
running during suspend.

The parameters of the screen off delay and the forced timeout
are configurable.

Signed-off-by: Alex Naidis <alex.naidis@linux.com>